Creating the user registration screen
The registration screen was modeled in Figma, and from there the component sizes were taken into account.

The registration screen has the title of the screen, a form to enter the name, email and password, as well as accept the terms and privacy policies, as well as the subscription for select product update. It also contains the buttons to complete the registration in the app or in Google.

For now, only the frontend part was done, for the display of the screen, however, in another corresponding issue the use of Redux and Hooks will be added, for the change of states of some components.

Co-authored-by: Alonso-AS <alfredoalonso_alcaraz@ucol.mx>